==> 03 json_to_r.Rmd <==
---
title: "03 json_to_r"
author: "JungHwan Yang"
date: "May 7, 2016"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

## Reading json file to R

This script reads a *.json file with political ideology labels (obamacare_labeled.json) and converts into R data frame.

```{r}
# install.packages("rjson"); library(rjson)
con <- file("D://obamacare_relabeled/obamacare_relabeled.json",open = "r")
obamacare<-vector("list")
i=0
while(T){
  i<-i+1
  temp<-readLines(con,1)
  if(length(temp)==0){
    print("reach the end")
    break
  } else
  obamacare[[i]]<-fromJSON(temp)
}
close(con)

rm(con); rm(i); rm(temp)
save(obamacare,file="D://obamacare_relabeled/obamacare_relabeled.RData")
```

## Extracting variables from a list object

The .RData is a large list that contains a number of list-in-list objects. The following code converts the subset of the list object into a R dataframe.

```{r}
load("~/obamacare_relabeled.RData") #modify according to a location of working directory

id_str <- sapply(obamacare, with, id_str)
user_ideology <- sapply(obamacare, with, ideology)
screen_name <- sapply(obamacare, with, user$screen_name)
name <- sapply(obamacare, with, user$name)
user_id_str <- sapply(obamacare, with, user$id_str)
text <- sapply(obamacare, with, text)
created_at <- sapply(obamacare, with, created_at)
verified <- sapply(obamacare, with, user$verified)
retweet_count <- sapply(obamacare, with, retweet_count)
statuses_count <- sapply(obamacare, with, user$statuses_count)
followers_count <- sapply(obamacare, with, user$followers_count)
friends_count <- sapply(obamacare, with, user$friends_count)
# some elements require to do as follows to handle a list
description <-
  sapply(obamacare, function(x) return(x$user$description))
description <- 
  sapply(description, function(x) ifelse(is.null(x), NA, x)) # list to vector
time_zone <-
  sapply(obamacare, function(x) return(x$user$time_zone))
time_zone <- 
  sapply(time_zone, function(x) ifelse(is.null(x), NA, x))
retweet_user <-
  sapply(obamacare, function(x) return(x$retweeted_status$user$screen_name))
retweet_user <-
  sapply(retweet_user, function(x) ifelse(is.null(x), NA, x))

obamacare_relabeled <-
  data.frame(id_str, user_ideology, screen_name, name, user_id_str, text,
             description, created_at, verified, statuses_count, retweet_count,
             retweet_user, time_zone, followers_count, friends_count)

rm(screen_name); rm(text); rm(retweet_count); rm(id_str); rm(created_at);
rm(verified); rm(user_id_str); rm(description); rm(followers_count);rm(name);
rm(retweet_user); rm(time_zone); rm(friends_count); rm(user_ideology);
rm(statuses_count)

# Remove white spaces in some variables due to format errors while saving .csv
# obamacare_relabeled$text <-
#   data.frame(gsub("\\s", " ", obamacare_relabeled$text))
# obamacare_relabeled$description <-
#   data.frame(gsub("\\s", " ", obamacare_relabeled$description))
# obamacare_relabeled$name <-
#   data.frame(gsub("\\s", " ", obamacare_relabeled$name))
obamacare_relabeled <-
  as.data.frame(apply(obamacare_relabeled, 2, function(x)gsub('\\s+', ' ',x)))

# Exporting dataframe in .csv
write.csv(obamacare_relabeled, file = "obamacare_relabeled.csv",
          row.names = F)
```
